Revert index/size types back to official signed types (#407)
* Revert index/size types back to official signed types

This change was made in c6e97cb8a0e27a44ee3ba678d713009cb50973d7 in attempt to reduce unnecessary casts & redundant checks, since most logic is basic unsigned logic.

Official logic continues to treat this as signed, so for simplicity's sake and to avoid unintentional issues with cases not refactored to behave correctly with unsigned logic,
we'll simply revert this so it's all signed again.

*Ideally* we'd just refactor affected logic to behave correctly with unsigned logic, but I fear this will cause more confusion with newer developers, introducing more bugs down the track,
so it's probably safer to just treat it all as signed.

Resolves #404

* Reformat N3UITrackBar.h

* Fix wonky for loop formatting

* Cleanup

+++ Client/N3Base/N3PMeshInstance.h
@@ -61,8 +61,8 @@ public:
 //	By : Ecli666 ( On 2002-08-06 오후 4:33:04 )
 //
 	void			PartialRender(int iCount, uint16_t* pIndices);
-	int				 GetIndexByiOrder(int iOrder);
-__Vector3		GetVertexByIndex(size_t iIndex);
+	int				GetIndexByiOrder(int iOrder);
+	__Vector3		GetVertexByIndex(int iIndex);
 //	~(By Ecli666 On 2002-08-06 오후 4:33:04 )
 };
